requirements.txt
# Core ML / data
torch==2.1.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2

# MLOps
mlflow==2.13.0
dvc==3.50.1

# Config
pyyaml==6.0.1

# Testing
pytest==8.2.0
pytest-cov==5.0.0
httpx==0.27.0

# Linting
flake8==7.0.0
